feat: customView 삭제
setSubviews 메서드를 외부에 노출해야하는 중대한 문제가 발생함
