+++ Discussion-Korea/Presentation/Common/ViewModelType.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-protocol ViewModelType {
+protocol ViewModelType: AnyObject {
 
     associatedtype Input
     associatedtype Output
